Updates Cronus
