+++ src/Elders.Cronus.Projections.Cassandra/Elders.Cronus.Projections.Cassandra.csproj
@@ -45,7 +45,7 @@
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
     </PackageReference>
-    <PackageReference Include="Cronus" Version="5.0.0-discovery0003" />
+    <PackageReference Include="Cronus" Version="5.0.0-discovery0005" />
     <PackageReference Include="CassandraCSharpDriver" Version="3.6.0" />
     <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="2.1.1" />
   </ItemGroup>
